+++ Modules/Intent.Modules.Java.Services.CRUD/ImplementationStrategies/JavaLombokAttributeExtensions.cs
@@ -1,5 +1,7 @@
 ﻿using Intent.Modelers.Domain.Api;
 using Intent.Modelers.Services.Api;
+using Intent.Modules.Common;
+using Intent.Modules.Common.Java;
 using Intent.Modules.Common.Templates;
 
 namespace Intent.Modules.Java.Services.CRUD.Decorators.ImplementationStrategies;
@@ -8,9 +10,9 @@ public static class JavaLombokAttributeExtensions
 {
     public static string Getter(this DTOFieldModel dtoField)
     {
-        if (dtoField.TypeReference.Element.Name == "bool")
+        if (dtoField.TypeReference?.HasBoolType() == true || dtoField.TypeReference?.HasJavaBooleanType() == true)
         {
-            return $"is{dtoField.Name.ToPascalCase().RemovePrefix("Is")}";
+            return $"is{dtoField.Name.ToPascalCase()}";
         }
         else
         {
@@ -20,7 +22,7 @@ public static class JavaLombokAttributeExtensions
 
     public static string Setter(this AttributeModel domainAttribute)
     {
-        if (domainAttribute.TypeReference.Element.Name == "bool")
+        if (domainAttribute.TypeReference?.HasBoolType() == true || domainAttribute.TypeReference?.HasJavaBooleanType() == true)
         {
             return $"set{domainAttribute.Name.ToPascalCase().RemovePrefix("Is")}";
         }